feat: Move ClientRepository to new folder structure
