+++ src/script/components/HistoryExport/HistoryExport.tsx
@@ -26,7 +26,7 @@ import {Button, ButtonVariant, FlexBox} from '@wireapp/react-ui-kit';
 import {LoadingBar} from 'Components/LoadingBar/LoadingBar';
 import {PrimaryModal} from 'Components/Modals/PrimaryModal';
 import {CancelError} from 'Repositories/backup/Error';
-import {ClientState} from 'src/script/client/ClientState';
+import {ClientState} from 'Repositories/client/ClientState';
 import {User} from 'src/script/entity/User';
 import {ContentState} from 'src/script/page/useAppState';
 import {t} from 'Util/LocalizerUtil';
