+++ src/script/components/Modals/LegalHoldModal/LegalHoldModal.test.tsx
@@ -22,11 +22,11 @@ import type {QualifiedId} from '@wireapp/api-client/lib/user/';
 
 import {useLegalHoldModalState} from 'Components/Modals/LegalHoldModal/LegalHoldModal.state';
 import {CallingRepository} from 'Repositories/calling/CallingRepository';
+import {ClientRepository} from 'Repositories/client';
 
 import {LegalHoldModal, LegalHoldModalType} from './LegalHoldModal';
 
 import {TestFactory} from '../../../../../test/helper/TestFactory';
-import {ClientRepository} from '../../../client';
 import {ConversationRepository} from '../../../conversation/ConversationRepository';
 import {MessageRepository} from '../../../conversation/MessageRepository';
 import {CryptographyRepository} from '../../../cryptography/CryptographyRepository';
