+++ src/script/components/UserDevices/components/DeviceList/DeviceList.tsx
@@ -19,11 +19,11 @@
 
 import cx from 'classnames';
 
+import type {ClientEntity} from 'Repositories/client/ClientEntity';
 import {WireIdentity} from 'src/script/E2EIdentity';
 import {useKoSubscribableChildren} from 'Util/ComponentUtil';
 import {t} from 'Util/LocalizerUtil';
 
-import type {ClientEntity} from '../../../../client/ClientEntity';
 import {Config} from '../../../../Config';
 import type {User} from '../../../../entity/User';
 import {DeviceCard} from '../DeviceCard';
