+++ src/script/entity/User/User.ts
@@ -27,12 +27,12 @@ import {Availability} from '@wireapp/protocol-messaging';
 import {WebAppEvents} from '@wireapp/webapp-events';
 
 import type {AssetRemoteData} from 'Repositories/assets/AssetRemoteData';
+import type {ClientEntity} from 'Repositories/client/ClientEntity';
 import {t} from 'Util/LocalizerUtil';
 import {clamp} from 'Util/NumberUtil';
 import {getFirstChar} from 'Util/StringUtil';
 import {TIME_IN_MILLIS} from 'Util/TimeUtil';
 
-import type {ClientEntity} from '../../client/ClientEntity';
 import {ACCENT_ID} from '../../Config';
 import {ConnectionEntity} from '../../connection/ConnectionEntity';
 import {ROLE as TEAM_ROLE} from '../../user/UserPermission';
