+++ src/script/main/app.ts
@@ -38,6 +38,8 @@ import {BackupRepository} from 'Repositories/backup/BackupRepository';
 import {BackupService} from 'Repositories/backup/BackupService';
 import {CacheRepository} from 'Repositories/cache/CacheRepository';
 import {CallingRepository} from 'Repositories/calling/CallingRepository';
+import {ClientRepository, ClientService} from 'Repositories/client';
+import {getClientMLSConfig} from 'Repositories/client/clientMLSConfig';
 import {initializeDataDog} from 'Util/DataDog';
 import {DebugUtil} from 'Util/DebugUtil';
 import {Environment} from 'Util/Environment';
@@ -51,8 +53,6 @@ import {AppInitializationStep, checkIndexedDb, InitializationEventLogger} from '
 import '../../style/default.less';
 import {SIGN_OUT_REASON} from '../auth/SignOutReason';
 import {URLParameter} from '../auth/URLParameter';
-import {ClientRepository, ClientService} from '../client';
-import {getClientMLSConfig} from '../client/clientMLSConfig';
 import {Configuration} from '../Config';
 import {ConnectionRepository} from '../connection/ConnectionRepository';
 import {ConnectionService} from '../connection/ConnectionService';
