+++ src/script/notification/PreferenceNotificationRepository.test.ts
@@ -23,11 +23,11 @@ import ko from 'knockout';
 
 import {WebAppEvents} from '@wireapp/webapp-events';
 
+import {ClientEntity} from 'Repositories/client/ClientEntity';
 import {Notification, PreferenceNotificationRepository} from 'src/script/notification/PreferenceNotificationRepository';
 import {PropertiesRepository} from 'src/script/properties/PropertiesRepository';
 import {createUuid} from 'Util/uuid';
 
-import {ClientEntity} from '../client/ClientEntity';
 import {User} from '../entity/User';
 
 describe('PreferenceNotificationRepository', () => {
