+++ src/script/page/AppLock/AppLock.test.tsx
@@ -24,11 +24,11 @@ import ko from 'knockout';
 
 import {WebAppEvents} from '@wireapp/webapp-events';
 
+import type {ClientRepository} from 'Repositories/client';
 import {createUuid} from 'Util/uuid';
 
 import {AppLock, APPLOCK_STATE} from './AppLock';
 
-import type {ClientRepository} from '../../client';
 import {TeamState} from '../../team/TeamState';
 import {AppLockRepository} from '../../user/AppLockRepository';
 import {AppLockState} from '../../user/AppLockState';
