+++ src/script/page/MainContent/MainContent.tsx
@@ -29,6 +29,7 @@ import {HistoryExport} from 'Components/HistoryExport';
 import {HistoryImport} from 'Components/HistoryImport';
 import * as Icon from 'Components/Icon';
 import {useLegalHoldModalState} from 'Components/Modals/LegalHoldModal/LegalHoldModal.state';
+import {ClientState} from 'Repositories/client/ClientState';
 import {User} from 'src/script/entity/User';
 import {useKoSubscribableChildren} from 'Util/ComponentUtil';
 import {t} from 'Util/LocalizerUtil';
@@ -41,7 +42,6 @@ import {AVPreferences} from './panels/preferences/AVPreferences';
 import {DevicesPreferences} from './panels/preferences/DevicesPreferences';
 import {OptionPreferences} from './panels/preferences/OptionPreferences';
 
-import {ClientState} from '../../client/ClientState';
 import {ConversationState} from '../../conversation/ConversationState';
 import {TeamState} from '../../team/TeamState';
 import {UserState} from '../../user/UserState';
