+++ src/script/page/MainContent/panels/preferences/DevicesPreferences/components/Device/Device.tsx
@@ -23,13 +23,13 @@ import {TabIndex} from '@wireapp/react-ui-kit/lib/types/enums';
 
 import {DeviceVerificationBadges} from 'Components/Badge';
 import * as Icon from 'Components/Icon';
+import {ClientEntity} from 'Repositories/client';
 import {WireIdentity} from 'src/script/E2EIdentity';
 import {useKoSubscribableChildren} from 'Util/ComponentUtil';
 import {handleKeyDown} from 'Util/KeyboardUtil';
 import {t} from 'Util/LocalizerUtil';
 import {splitFingerprint} from 'Util/StringUtil';
 
-import {ClientEntity} from '../../../../../../../client';
 import {FormattedId} from '../FormattedId';
 
 interface DeviceProps {
