+++ src/script/page/MainContent/panels/preferences/accountPreferences/LogoutSection.tsx
@@ -23,9 +23,8 @@ import {TabIndex} from '@wireapp/react-ui-kit/lib/types/enums';
 
 import {Button, ButtonVariant} from '@wireapp/react-ui-kit';
 
+import type {ClientRepository} from 'Repositories/client';
 import {t} from 'Util/LocalizerUtil';
-
-import type {ClientRepository} from '../../../../../client';
 
 interface LogoutSectionProps {
   clientRepository: ClientRepository;
