+++ src/script/repositories/client/ClientRepository.ts
@@ -29,6 +29,7 @@ import {container} from 'tsyringe';
 import {Runtime} from '@wireapp/commons';
 import {WebAppEvents} from '@wireapp/webapp-events';
 
+import {PrimaryModal} from 'Components/Modals/PrimaryModal';
 import {t} from 'Util/LocalizerUtil';
 import {getLogger, Logger} from 'Util/Logger';
 import {matchQualifiedIds} from 'Util/QualifiedId';
@@ -41,14 +42,13 @@ import type {ClientService} from './ClientService';
 import {ClientState} from './ClientState';
 import {isClientMLSCapable, wasClientActiveWithinLast4Weeks} from './ClientUtils';
 
-import {SIGN_OUT_REASON} from '../auth/SignOutReason';
-import {PrimaryModal} from '../components/Modals/PrimaryModal';
-import type {CryptographyRepository} from '../cryptography/CryptographyRepository';
-import type {User} from '../entity/User';
-import {ClientError} from '../error/ClientError';
-import {Core} from '../service/CoreSingleton';
-import {ClientRecord} from '../storage';
-import {StorageKey} from '../storage/StorageKey';
+import {SIGN_OUT_REASON} from '../../auth/SignOutReason';
+import type {CryptographyRepository} from '../../cryptography/CryptographyRepository';
+import type {User} from '../../entity/User';
+import {ClientError} from '../../error/ClientError';
+import {Core} from '../../service/CoreSingleton';
+import {ClientRecord} from '../../storage';
+import {StorageKey} from '../../storage/StorageKey';
 
 export type UserClientEntityMap = {[userId: string]: ClientEntity[]};
 export type QualifiedUserClientEntityMap = {[domain: string]: UserClientEntityMap};
