+++ src/script/repositories/client/ClientUtils.ts
