+++ src/script/repositories/client/clientMLSConfig.ts
@@ -19,9 +19,9 @@
 
 import {FeatureList} from '@wireapp/api-client/lib/team';
 
-import {Config} from '../Config';
-import {getE2EIConfig} from '../page/components/FeatureConfigChange/FeatureConfigChangeHandler/Features/E2EIdentity';
-import {getMLSConfig} from '../page/components/FeatureConfigChange/FeatureConfigChangeHandler/Features/MLS';
+import {Config} from '../../Config';
+import {getE2EIConfig} from '../../page/components/FeatureConfigChange/FeatureConfigChangeHandler/Features/E2EIdentity';
+import {getMLSConfig} from '../../page/components/FeatureConfigChange/FeatureConfigChangeHandler/Features/MLS';
 
 export function getClientMLSConfig(teamFeatures: FeatureList) {
   const keyingMaterialUpdateThreshold = Config.getConfig().FEATURE.MLS_CONFIG_KEYING_MATERIAL_UPDATE_THRESHOLD;
