+++ src/script/user/UserRepository.test.ts
@@ -29,6 +29,8 @@ import {Availability} from '@wireapp/protocol-messaging';
 import {WebAppEvents} from '@wireapp/webapp-events';
 
 import {AssetRepository} from 'Repositories/assets/AssetRepository';
+import {ClientRepository} from 'Repositories/client';
+import {ClientMapper} from 'Repositories/client/ClientMapper';
 import {entities} from 'test/api/payloads';
 import {TestFactory} from 'test/helper/TestFactory';
 import {generateAPIUser} from 'test/helper/UserGenerator';
@@ -39,8 +41,6 @@ import {UserRepository} from './UserRepository';
 import {UserService} from './UserService';
 import {UserState} from './UserState';
 
-import {ClientRepository} from '../client';
-import {ClientMapper} from '../client/ClientMapper';
 import {ConnectionEntity} from '../connection/ConnectionEntity';
 import {User} from '../entity/User';
 import {EventRepository} from '../event/EventRepository';
