+++ src/script/user/UserRepository.ts
@@ -47,6 +47,9 @@ import {WebAppEvents} from '@wireapp/webapp-events';
 import {useLegalHoldModalState} from 'Components/Modals/LegalHoldModal/LegalHoldModal.state';
 import {mapProfileAssetsV1} from 'Repositories/assets/AssetMapper';
 import {AssetRepository} from 'Repositories/assets/AssetRepository';
+import type {ClientRepository, QualifiedUserClientEntityMap} from 'Repositories/client';
+import {ClientEntity} from 'Repositories/client/ClientEntity';
+import {ClientMapper} from 'Repositories/client/ClientMapper';
 import {chunk, partition} from 'Util/ArrayUtil';
 import {t} from 'Util/LocalizerUtil';
 import {getLogger, Logger} from 'Util/Logger';
@@ -61,9 +64,6 @@ import type {UserService} from './UserService';
 import {UserState} from './UserState';
 
 import {SIGN_OUT_REASON} from '../auth/SignOutReason';
-import type {ClientRepository, QualifiedUserClientEntityMap} from '../client';
-import {ClientEntity} from '../client/ClientEntity';
-import {ClientMapper} from '../client/ClientMapper';
 import {Config} from '../Config';
 import type {ConnectionEntity} from '../connection/ConnectionEntity';
 import {flattenUserClientsQualifiedIds} from '../conversation/userClientsUtils';
