+++ src/script/util/DebugUtil.ts
@@ -41,14 +41,14 @@ import {showAppNotification} from 'Components/AppNotification';
 import {CallingRepository} from 'Repositories/calling/CallingRepository';
 import {CallState} from 'Repositories/calling/CallState';
 import {Participant} from 'Repositories/calling/Participant';
+import {ClientRepository} from 'Repositories/client';
+import {ClientState} from 'Repositories/client/ClientState';
 import {getLogger, Logger} from 'Util/Logger';
 
 import {KEY} from './KeyboardUtil';
 import {TIME_IN_MILLIS} from './TimeUtil';
 import {createUuid} from './uuid';
 
-import {ClientRepository} from '../client';
-import {ClientState} from '../client/ClientState';
 import {ConnectionRepository} from '../connection/ConnectionRepository';
 import {ConversationRepository} from '../conversation/ConversationRepository';
 import {isMLSCapableConversation} from '../conversation/ConversationSelectors';
