+++ src/script/view_model/MainViewModel.ts
@@ -23,13 +23,13 @@ import type {AssetRepository} from 'Repositories/assets/AssetRepository';
 import type {AudioRepository} from 'Repositories/audio/AudioRepository';
 import type {BackupRepository} from 'Repositories/backup/BackupRepository';
 import type {CallingRepository} from 'Repositories/calling/CallingRepository';
+import type {ClientRepository} from 'Repositories/client';
 
 import {ActionsViewModel} from './ActionsViewModel';
 import {CallingViewModel} from './CallingViewModel';
 import {ContentViewModel} from './ContentViewModel';
 import {ListViewModel} from './ListViewModel';
 
-import type {ClientRepository} from '../client';
 import type {ConnectionRepository} from '../connection/ConnectionRepository';
 import type {ConversationRepository} from '../conversation/ConversationRepository';
 import type {MessageRepository} from '../conversation/MessageRepository';
