+++ test/helper/TestFactory.js
@@ -31,10 +31,10 @@ import {container} from 'tsyringe';
 import {AssetRepository} from 'Repositories/assets/AssetRepository';
 import {AudioRepository} from 'Repositories/audio/AudioRepository';
 import {CallingRepository} from 'Repositories/calling/CallingRepository';
-import {ClientEntity} from 'src/script/client/ClientEntity';
-import {ClientRepository} from 'src/script/client/ClientRepository';
-import {ClientService} from 'src/script/client/ClientService';
-import {ClientState} from 'src/script/client/ClientState';
+import {ClientEntity} from 'Repositories/client/ClientEntity';
+import {ClientRepository} from 'Repositories/client/ClientRepository';
+import {ClientService} from 'Repositories/client/ClientService';
+import {ClientState} from 'Repositories/client/ClientState';
 import {ConnectionRepository} from 'src/script/connection/ConnectionRepository';
 import {ConnectionService} from 'src/script/connection/ConnectionService';
 import {ConversationRepository} from 'src/script/conversation/ConversationRepository';
